add rock-import
This is meant as starting point for a tool which:
imports logfiles by converting/updating them to a version
which can be replayed into the current system. It also has
some functionality for handling logfiles like building
an index for all logfiles, etc

At the moment only building an index for a hole folder is implemented!