add a generic post-processing class
The class currently only implements selection of streams and regions
within the streams, and implements the output parts (i.e. generate
log files which are made of the selected streams).

The obvious next steps are (1) to create a UI and a command line tool
to build generic processing chains and (2) to add filtering capabilities
and integrate the possibility of running the datatype upgrade handlers
as well as export handlers (for instance write image streams as sets of
standard image files)